Make Location a top-level class.
